+++ apps/docs/src/styles/v2.css
@@ -1,5 +1,5 @@
 /* ============================================================
-   Durabull /v2 landing — "Engineering Grade" design system
+   Durabull home landing — "Engineering Grade" design system
    Scoped under .v2 so it never leaks into the rest of the site.
    White/black section contrast, architectural grid, restrained
    ember accent. Light-first; sections opt into .v2-dark.
